requirements.txt
asgiref==3.5.2
black==22.8.0
click==8.1.3
Django==4.1.1
django-environ==0.9.0
django-taggit==3.0.0
Markdown==3.4.1
mypy-extensions==0.4.3
pathspec==0.10.1
platformdirs==2.5.2
psycopg2-binary==2.9.3
sqlparse==0.4.2
tomli==2.0.1
